In ir_ntzl, always use the 64 bit version on Windows
When compiling on Windows with clang-cl, __has_builtin(__builtin_ctzl)
will evaluate to true, but a long on Windows is only 32 bits, resulting
in an incorrect implementation. Move the _WIN64 block ahead of the
general block so that it gets the 64 bit version.

Fixes https://github.com/dstogov/ir/issues/101.