Fix php/php-src#17966: Symfony JIT 1205 assertion failure
In this example code,
`ir_split_partially_dead_node()` tries to create a clone of an IR_COPY
instruction, but that instruction has an extra data op in op2.
IR tries to add uses to op2 but it's not actually a real ir_ref.
In this case, op2 is the number 1; if it were an ir_ref it would
correspond to the IR_START node, which seems wrong.
This patch adds an extra check to see if it's a real input.